+++ src/patch/dll_collision_viewer.zig
@@ -279,6 +279,7 @@ const QuickRaceMenu = extern struct {
     }
 
     fn open() void {
+        gf.ASettingSaveAuto();
         if (!gf.GFreezeOn()) return;
         rs.swrSound_PlaySound(78, 6, 0.25, 1.0, 0);
         data.idx = 0;
@@ -286,6 +287,7 @@ const QuickRaceMenu = extern struct {
     }
 
     fn close() void {
+        gf.ASettingSaveAuto();
         if (!gf.GFreezeOff()) return;
         rs.swrSound_PlaySound(77, 6, 0.25, 1.0, 0);
         menu_active = false;